requirements.txt
anthropic>=0.18.0
requests>=2.31.0
jsonschema>=4.0.0
python-dotenv>=1.0.0
